+++ sensors/nivel/constants.py
@@ -1,0 +1,4 @@
+MODEL_TYPE_CHOICE = (
+    (201, "Leica Nivel"),
+    (202, "Seika N10"),
+)