+++ tachy/lib/constants.py
@@ -1,5 +1,0 @@
-MODEL_TYPE_CHOICE = (
-    (11, "Leica 1100-Series"),
-    (12, "Leica 1200-Series"),
-    (15, "Leica TS15"),
-)