+++ CHANGES.md
@@ -1,3 +1,18 @@
+
+# dimosy django 2016-01-20
+
+- working on structure
+- added app 'sensors'
+- moved tachy under sensors
+- added app 'sensors.nivel'
+
+
+# dimosy django 2016-01-11
+
+- fixed issue #17
+- added "active" to time window class
+
+
 # dimosy django 2015-12-21
 
 - added CHANGES.md
@@ -8,8 +23,3 @@
 - unicode for models represantation
 - added setup to urlpatterns
 
-
-# dimosy django 2016-01-11
-
-- fixed issue #17
-- added "active" to time window class